requirements.txt
django==4.2.*
environs[django]==11
pillow==11
django-admin-sortable2==2.2.3
django-tinymce==4.1.0
requests==2.32.*
